Mask Detection Using Python
This program is made in python. It using Machine Learning for the detection of Face mask. Here OpenCV and SKlearn modules are used for the program 
In order to run the program just run all the tabs(try using jupyter notebook). CLICK ESC to terminate the program.
Note: Run the program in the light area, not in dark.